requirements.txt
aiogram==3.4.1
# aiogram-redis==1.0.0  # Примітка:Якщо не знайде aiogram-redis, можна використовувати стандартний redis і FSMStorage з aiogram.
python-dotenv==1.0.1
redis==6.2.0
flask==3.0.3